MB-41321: Collection start-seqno can never be negative
Negative sequence numbers are reserved for temporary hash-table items
only, the recorded start-sequence number of a collection is always
positive.

Change start sequence number from signed to unsigned.

This is re-factoring to simplify the main fix for MB-41321

Reviewed-on: http://review.couchbase.org/c/kv_engine/+/136539
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/collections/vbucket_manifest.cc
@@ -544,7 +544,7 @@ bool Manifest::isLogicallyDeleted(const container::const_iterator entry,
 
     // seqno >= 0 (so temp items etc... are ok) AND the seqno is below the
     // collection start (start is set on creation and moves with flush)
-    return seqno >= 0 && seqno < entry->second.getStartSeqno();
+    return seqno >= 0 && (uint64_t(seqno) < entry->second.getStartSeqno());
 }
 
 void Manifest::processExpiryTime(const container::const_iterator entry,
